feat: extract census and openrouter helpers
